requirements.txt
Authlib==1.3.1
bcrypt==4.1.3
blinker==1.8.2
cachelib==0.13.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-OAuthlib==0.9.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
google-auth==2.30.0
google-auth-oauthlib==1.2.0
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector==2.2.9
mysql-connector-python==8.4.0
oauthlib==2.1.0
pillow==10.3.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==1.1.0
rsa==4.9
SQLAlchemy==2.0.30
typing_extensions==4.12.0
urllib3==2.2.1
Werkzeug==3.0.3
WTForms==3.1.2